@@ -27,7 +27,7 @@
 		"prettier": "^3.3.0",
 		"prettier-plugin-svelte": "^3.1.2",
 		"prettier-plugin-tailwindcss": "^0.6.1",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0-next.214",
 		"svelte-check": "^3.6.0",
 		"tailwindcss": "^3.4.3",
 		"tslib": "^2.4.1",
@@ -37,6 +37,7 @@
 	},
 	"type": "module",
 	"dependencies": {
+		"@fortawesome/fontawesome-free": "^6.6.0",
 		"@sveltejs/adapter-node": "^5.0.1"
 	}
 }
