@@ -1,6 +1,7 @@
 <script lang="ts">
 	import { onNavigate } from '$app/navigation';
 	import './styles.css';
+	import NavItems from '$lib/NavItems.svelte';
 
 	onNavigate((navigation) => {
 		// @ts-ignore
@@ -16,80 +17,42 @@
 	});
 </script>
 
-<div class="app bg-base-300 flex min-h-dvh flex-col">
-	<div class="bg-base-200 w-full">
-		<div class="navbar mx-auto max-w-screen-2xl">
-			<div class="flex-1">
-				<a href="/" class="btn btn-ghost text-xl">GW2Tools</a>
+<div class="app flex min-h-dvh flex-col">
+	<div class="bg-base-200 flex flex-row justify-center">
+		<div class="navbar max-w-screen-2xl">
+			<div class="navbar-start">
+				<div class="dropdown">
+					<div tabindex="0" role="button" class="btn btn-ghost lg:hidden">
+						<i class="fa-solid fa-bars"></i>
+					</div>
+					<ul
+						class="menu menu-sm dropdown-content bg-base-100 rounded-box top-10 z-[2] mt-3 w-40 gap-y-2 p-2 shadow"
+					>
+						<NavItems />
+					</ul>
+				</div>
+				<a class="btn btn-ghost text-xl" href="/">GW2Tools</a>
 			</div>
-			<div class="flex-none">
+			<div class="navbar-center hidden lg:flex">
 				<ul class="menu menu-horizontal px-1">
-					<li class="hidden sm:block"><a href="/"><i class="fa-solid fa-house"></i>Home</a></li>
-					<li class="hidden sm:block">
-						<a href="/reset"><i class="fa-solid fa-arrows-rotate"></i>Reset</a>
-					</li>
-					<li class="hidden sm:block">
-						<a href="/magicfind"><i class="fa-solid fa-clover"></i>Magic Find</a>
-					</li>
-					<li class="hidden sm:block">
-						<a href="/ecto"><i class="fa-solid fa-screwdriver-wrench"></i>Ecto Salvage</a>
-					</li>
-					<li>
-						<label class="swap swap-rotate">
-							<!-- this hidden checkbox controls the state -->
-							<input type="checkbox" class="theme-controller" value="light" />
-
-							<!-- sun icon -->
-							<svg
-								class="swap-off size-5 fill-current"
-								xmlns="http://www.w3.org/2000/svg"
-								viewBox="0 0 24 24"
-								><path
-									d="M5.64,17l-.71.71a1,1,0,0,0,0,1.41,1,1,0,0,0,1.41,0l.71-.71A1,1,0,0,0,5.64,17ZM5,12a1,1,0,0,0-1-1H3a1,1,0,0,0,0,2H4A1,1,0,0,0,5,12Zm7-7a1,1,0,0,0,1-1V3a1,1,0,0,0-2,0V4A1,1,0,0,0,12,5ZM5.64,7.05a1,1,0,0,0,.7.29,1,1,0,0,0,.71-.29,1,1,0,0,0,0-1.41l-.71-.71A1,1,0,0,0,4.93,6.34Zm12,.29a1,1,0,0,0,.7-.29l.71-.71a1,1,0,1,0-1.41-1.41L17,5.64a1,1,0,0,0,0,1.41A1,1,0,0,0,17.66,7.34ZM21,11H20a1,1,0,0,0,0,2h1a1,1,0,0,0,0-2Zm-9,8a1,1,0,0,0-1,1v1a1,1,0,0,0,2,0V20A1,1,0,0,0,12,19ZM18.36,17A1,1,0,0,0,17,18.36l.71.71a1,1,0,0,0,1.41,0,1,1,0,0,0,0-1.41ZM12,6.5A5.5,5.5,0,1,0,17.5,12,5.51,5.51,0,0,0,12,6.5Zm0,9A3.5,3.5,0,1,1,15.5,12,3.5,3.5,0,0,1,12,15.5Z"
-								/></svg
-							>
-							<!-- moon icon -->
-							<svg
-								class="swap-on size-5 fill-current"
-								xmlns="http://www.w3.org/2000/svg"
-								viewBox="0 0 24 24"
-								><path
-									d="M21.64,13a1,1,0,0,0-1.05-.14,8.05,8.05,0,0,1-3.37.73A8.15,8.15,0,0,1,9.08,5.49a8.59,8.59,0,0,1,.25-2A1,1,0,0,0,8,2.36,10.14,10.14,0,1,0,22,14.05,1,1,0,0,0,21.64,13Zm-9.5,6.69A8.14,8.14,0,0,1,7.08,5.22v.27A10.15,10.15,0,0,0,17.22,15.63a9.79,9.79,0,0,0,2.1-.22A8.11,8.11,0,0,1,12.14,19.73Z"
-								/></svg
-							>
-						</label>
-					</li>
-					<li class="sm:hidden">
-						<details>
-							<summary><i class="fa-solid fa-bars"></i>Menu</summary>
-							<ul class="menu bg-base-200 z-50 rounded-t-none p-2">
-								<li><a href="/"><i class="fa-solid fa-house"></i>Home</a></li>
-								<li><a href="/reset"><i class="fa-solid fa-arrows-rotate"></i>Reset</a></li>
-								<li><a href="/magicfind"><i class="fa-solid fa-clover"></i>Magic Find</a></li>
-								<li>
-									<a href="/ecto"><i class="fa-solid fa-screwdriver-wrench"></i>Ecto Salvage</a>
-								</li>
-							</ul>
-						</details>
-					</li>
+					<NavItems />
 				</ul>
 			</div>
+			<div class="navbar-end"></div>
 		</div>
 	</div>
 
-	<main class="mx-auto w-full max-w-screen-2xl flex-grow p-2">
+	<main class="bg-base-300 mx-auto w-full grow pb-4">
 		<slot />
 	</main>
 
-	<footer class="footer flex justify-center p-2">
-		<div class="flex w-full flex-row items-center gap-8">
-			<div class="flex w-full justify-center">
-				<span class="text-xs">© henahax.de 2024</span>
-			</div>
-			<a
-				class="btn btn-ghost btn-sm flex w-fit justify-end text-xs"
-				href="https://github.com/Henahax/GW2Tools">Sourcecode</a
-			>
+	<footer class="footer bg-base-200 flex items-center p-2 text-xs">
+		<div class="grow"></div>
+		<div class="w-fit">© henahax.de 2024</div>
+		<div class="flex grow justify-end">
+			<a class="btn btn-ghost btn-xs" href="https://github.com/Henahax/GW2Tools">
+				<i class="fa-brands fa-github"></i>Sourcecode
+			</a>
 		</div>
 	</footer>
 </div>
